refactor(Calculator): Tokenizer 유틸 객체로 변경
Tokenizer 유틸 객체로 변경
- 기존에 인스턴스 생성해서 사용하던 Tokenizer를 static으로 변경
- 변경하면서 util 패키지 내부로 객체 위치 변경
