+++ src/main/java/calculator/model/enums/ParsingPattern.java
@@ -1,4 +1,4 @@
-package calculator.model.token;
+package calculator.model.enums;
 
 public enum ParsingPattern {
     DEFAULT_DELIMITER_PATTERN(",|:"),
