+++ src/main/java/calculator/view/InputView.java
@@ -1,6 +1,6 @@
 package calculator.view;
 
-import static calculator.view.ViewMessage.NUMBERS_INPUT_MESSAGE;
+import static calculator.view.enums.ViewMessage.NUMBERS_INPUT_MESSAGE;
 import static camp.nextstep.edu.missionutils.Console.readLine;
 
 public class InputView {
