+++ src/main/java/calculator/view/OutputView.java
@@ -1,6 +1,6 @@
 package calculator.view;
 
-import static calculator.view.ViewMessage.RESULT_MESSAGE;
+import static calculator.view.enums.ViewMessage.RESULT_MESSAGE;
 
 public class OutputView {
     public void printResult(int result) {
