+++ src/main/java/calculator/view/enums/ViewMessage.java
@@ -1,4 +1,4 @@
-package calculator.view;
+package calculator.view.enums;
 
 public enum ViewMessage {
     NUMBERS_INPUT_MESSAGE("덧셈할 문자열을 입력해 주세요."),
